Deleted bipoler from name.  Deleted commented out pdb line.  Refactored unnecessary else statement. Moved test and strategy into existing titfortat files.  Removed automodule for deleted rand_titfortat file.
