+++ docs/reference/all_strategies.rst
@@ -138,9 +138,6 @@ Here are the docstrings of all the strategies in the library.
 .. automodule:: axelrod.strategies.rand
    :members:
    :undoc-members:
-.. automodule:: axelrod.strategies.rand_titfortat
-   :members:
-   :undoc-members:
 .. automodule:: axelrod.strategies.resurrection
    :members:
    :undoc-members:
